refactor: move Footer component to components directory
