+++ src/pages/Home.jsx
@@ -4,7 +4,7 @@ import OurStory from '../components/sections/OurStory.jsx';
 import HowWeWork from '../components/sections/HowWeWork.jsx';
 import Projects from '../components/sections/Projects.jsx';
 import JoinUs from '../components/sections/JoinUs.jsx';
-import Footer from '../components/sections/Footer.jsx';
+import Footer from '../components/Footer.jsx';
 
 function Home() {
   return (
